# docs/swagger.yaml
basePath: /work-order-console/api/v1
definitions:
  request.LoginRequest:
    properties:
      password:
        description: 密码， 使用公钥加密后传输
        type: string
      username:
        description: 用户名
        type: string
    type: object
  request.UserAddRequest:
    properties:
      password:
        description: 密码
        type: string
      role:
        description: 角色
        type: string
      username:
        description: 用户名
        type: string
    type: object
  request.UserModifyRequest:
    properties:
      role:
        description: 角色
        type: string
    type: object
  response.PubKeyResponse:
    properties:
      pubKey:
        description: 公钥， X509 HEX 格式
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: 提示码
        type: string
      data:
        description: 数据
      message:
        description: 提示信息
        type: string
      requestId:
        description: 请求id
        type: string
    type: object
  response.UserDetailResponse:
    properties:
      id:
        description: 用户id
        type: string
      role:
        description: 用户角色
        type: string
      username:
        description: 用户名
        type: string
    type: object
  response.UserListResponse:
    properties:
      total:
        description: 总数
        type: integer
      users:
        description: 用户数组
        items:
          $ref: '#/definitions/response.UserDetailResponse'
        type: array
    type: object
info:
  contact: {}
  description: 开发人员使用的后台运维系统
  title: 工单控制台
  version: 1.0.0
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 登陆
      parameters:
      - description: body
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.Response'
      summary: 登陆
      tags:
      - 登陆管理
  /logout:
    get:
      consumes:
      - application/json
      description: 退出登陆
      parameters:
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.Response'
      summary: 退出登陆
      tags:
      - 登陆管理
  /system/pubKey:
    get:
      consumes:
      - application/json
      description: 获取公钥
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.PubKeyResponse'
      summary: 获取公钥
      tags:
      - 系统关系
  /user:
    get:
      consumes:
      - application/json
      description: 用户列表
      parameters:
      - description: pageIndex
        in: query
        name: pageIndex
        required: true
        type: string
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: string
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 角色
        in: query
        name: role
        required: true
        type: string
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.UserListResponse'
      summary: 用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: body
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/request.UserAddRequest'
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建用户
      tags:
      - 用户管理
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 用户详情
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.UserDetailResponse'
      summary: 用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 修改用户
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/request.UserModifyRequest'
      - description: Token
        in: header
        name: Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改用户
      tags:
      - 用户管理
swagger: "2.0"